Add a script to migrate existing LSTM model from Keras 2 to Keras 3
This script converts action_recognition_model.h5 to a valid Keras 3
model, based on the structure defined in TrainingLSTM.py.

The new model was generated using the command:
`uv run .\Model\Test\MigrateModel.py`. This change does not delete
the original model, but it can be removed once the whole project is
migrated to Keras 3 and contributors verify that there is no changes in
the model behavior or training flows.